(Refactor) int 변수 자료형 변경
메모리 최적화를 위한 sbyte, short 사용

+++ Assets/Scripts/Bullet.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class Bullet : MonoBehaviour
 {
     public enum BulletType { Melee, Range }
-    public int damage;
-    public int penetrate;
+    public short damage;
+    public sbyte penetrate;
     public float speed;
 
     Rigidbody2D rigid;
